Added Test Movement Script || Requires much refinement
So far walking, running, crouching with roll and a very basic jump have been added. Adjustments need to be made to the rolling ability as it currently moves the player at an instant. Same can be said for the jump, however variations of the jump will also need to be added.

Movement modes (gun out/holstered) will be added as soon as the basic movement has been achieved.

Code will also require a clean-up once the basics have been done.