Refined Test Movement Script || Still Requires Work
This lesson I was able to figure out the solution to the problem of "Why isn't AddForce working?". Currently walking and jumping uses AddForce, but not running or rolling. There are now three jumping styles, although they still require major tinkering.

Player direction has now been established, the placeholder player has a "nose" to distinguish where it is looking. With the addition of player direction, walking and running need a slight rework to reflect the player moving where it is looking.

With how jumping currently works, if the player is in mid air long enough they are able to double jump. I will fix this ASAP.