Camera Switch Added, Continued Work Refining Movement || Movement Requires A Lot of Work Still
Lots of changes made to the movement script across the weekend. Player walking and running has been changed back to translate for the meantime as movement speed appears to be much more consistent when implemented this way.

The player cannot change the direction they are facing during the time they are in midair, nor can they instantly move as they change walking/running direction with the newly applied "turn delay". The player climbing system is still a work in progress as I still have to figure out the implementation of that system.

It also appears that values in the inspector behave differently for the script in either the school computer copy of Unity and my home copy, or perhaps it is my imagination.

-------------------------

ALSO now that I remember, the player is now unable to affect their movement trajectory while in midair.