GitHub Commitment Test || Scene Assets Not Pushing Properly?
This GitHub commit is to test whether the newly created PlayerTest and LedgePosition prefabs commit properly.

This is in response to an event that happened on Saturday afternoon where I found that in my scene, certain assets such as scripts and prefabs were not being found for objects such as the floor detector script on the player's feet, the LedgePosition prefabs all over the map and the materials on the scene transition objects and player ledge detector.

Here's hoping there are no issues later this morning.