Work On Enemy AI
The rifle-armed AI are now capable of crouching when the player crouches.

The enemy crawlers now apply knockback to the player when they get too close. At the same time that knockback is applied, the player is also sent into a grace period where knockback (and later on damage) does not apply to the player.