Player Knockback, Robot Ledge Detection, AI Combat
When the player takes a hit from the enemy, they recieve knockback.

Both enemy variations can now attack the player and deal damage to the player's shield, although currently the damage dealt is not consistent and there is no death condition set on to the player.

The robot enemy has gained the ability to detect ledges. Now, should we ever decide to have robot enemies in locations where they could potentially fall of ledges, there is a system to prevent robots falling.