Camera Switches and Ledges Grabs Placed In Level 1 || Orientation Fix
Camera switches and ledge grab locations have been placed in the level 1 scene, and the orientation of the level has also been fixed.

NOTE: For the orientation fix, to avoid breaking anything already in the scene, I have in fact created a new scene. Pretty much a duplicate of the old scene but with the orientation fix. This orientation fix was required as the animations for ledge climbing would have been affected by the wrong-facing orientation of the old scene.