Fixed Enemy AI Scripts || Slightly Refined Ledge Collision
Now enemy humanoids deal consistent damage to the player. This is not the case for enemy crawlers as of yet. Enemy crawlers now cannot be hit by the player until the crawler pulls out their taser thingy.

Ledge collision has been slightly altered so that the player cannot grab on to a ledge unless they are facing the right direction. Although ledges still need further altering, this fix should stop players from grabbing the right-end of ledge prefabs when left-end is the end that is exposed.