Implemented Most Player Animations
As in the title, most of the player animations have been implemented. The animations are a bit glitchy, likely due to some code areas of the code that I have failed to update, however this commit is here for those interested in viewing my progress so far, with how close the due date is.